services/ml/pyproject.toml
[build-system]
requires = ["setuptools>=68,<82", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foodsea-ml-service"
version = "0.1.0"
description = "FoodSea ml-service (gRPC semantic analog search)"
readme = "TUTORIAL.md"
requires-python = ">=3.11,<4"
dependencies = [
  "grpcio==1.62.2",
  "grpcio-tools==1.62.2",
  "sentence-transformers==3.0.1",
  "scikit-learn==1.5.1",
  "numpy==1.26.4",
  "protobuf==4.25.3",
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]
